Move inner classes of Command and Task outside
