+++ src/main/java/duke/command/EmptyCommand.java
@@ -1,0 +1,14 @@
+package duke.command;
+
+import duke.task.TaskList;
+
+public class EmptyCommand extends Command {
+    public EmptyCommand(String[] args) {
+        super(CommandType.EMPTY, args);
+    }
+
+    @Override
+    public void runSpecialTask(TaskList tasks) {
+        // does nothing
+    }
+}
