+++ src/main/java/duke/task/TodoTask.java
@@ -1,0 +1,7 @@
+package duke.task;
+
+public class TodoTask extends Task {
+    public TodoTask(String[] args) {
+        super(args[0], 'T');
+    }
+}
