+++ src/main/java/duke/util/Parser.java
@@ -1,0 +1,5 @@
+package duke.util;
+
+public class Parser {
+    
+}
